Extract block related code to block.js
